projeto concluido

+++ excluir.php
@@ -11,17 +11,26 @@ $consulta->bindParam(':cd', $cd, PDO::PARAM_INT);
 $consulta->execute();
 $exibe = $consulta->fetch(PDO::FETCH_ASSOC);
 
-
 $excluir = $cn->prepare("DELETE FROM tbl_produto WHERE cd_produto = :cd");
 $excluir->bindParam(':cd', $cd, PDO::PARAM_INT);
-$excluir->execute();
 
-$foto1 = $exibe['img_produto'];  
+if ($excluir->execute()) {
+    $foto1 = $exibe['img_produto'];
 
-if (!empty($foto1)) {  
-    unlink($pasta . $foto1);
+    if (!empty($foto1)) {  
+        unlink($pasta . $foto1);
+    }
+    
+    // Redirecionar com JavaScript após a exclusão bem-sucedida
+    echo '<script>
+            alert("Produto excluído com sucesso!");
+            window.location.href = "lista.php";
+          </script>';
+} else {
+    // Redirecionar com JavaScript em caso de erro na exclusão
+    echo '<script>
+            alert("Erro ao excluir o produto. Por favor, tente novamente.");
+            window.location.href = "lista.php";
+          </script>';
 }
-
-
-header('location: lista.php');
 ?>
